requirements.txt
# Core dependencies
streamlit==1.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
aiodns==3.0.0
aiofiles==23.1.0
nest-asyncio==1.5.7
asyncio==3.4.3
websockets>=9.0,<12.0

# Solana specific
solana>=0.30.2
anchorpy>=0.18.0
solders>=0.18.0

# Data processing and analysis
pandas>=2.1.4
numpy>=1.26.2
pyarrow==12.0.1
ta==0.10.2
scipy>=1.11.4

# Database
sqlalchemy==2.0.19
alembic==1.11.1
aiosqlite==0.19.0

# Visualization
plotly==5.15.0
matplotlib>=3.8.2
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Logging and monitoring
python-json-logger==2.0.7

# Additional dependencies
python-telegram-bot>=20.7
requests>=2.31.0
python-binance>=1.0.19
ccxt>=4.1.87
